+++ lib/src/internal/wiggly_defaults.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class WigglyDefaults {
+  static const Color loaderProgressColor = Color(0xFF3B82F6);
+  static const Color loaderTrackColor = Color(0xFFE5E7EB);
+
+  static const Color linearProgressColor = Color(0xFF3B82F6);
+  static const Color linearTrackColor = Color(0xFFE5E7EB);
+
+  static const Color refreshProgressColor = Color(0xFF3B89F6);
+  static const Color refreshTrackColor = Color(0xFFE5E7EB);
+  static const Color refreshBackgroundColor = Colors.white;
+}
+